+++ classifier/classifier/training.py
@@ -1,6 +1,6 @@
 import lxml.html  # @UnresolvedImport
 import requests
-from cms.scraper import extract_words
+from scraper.scraper import extract_words
 import pickle
 import sys
 import os
